add params.  move _as_operator to relaxation. limit circular imports
