+++ setup.cfg
@@ -93,8 +93,7 @@ ignore-names =
 
 [pylint:MESSAGES CONTROL]
 disable=invalid-name,too-many-lines,too-many-locals,too-many-branches,too-many-statements,too-many-arguments,too-few-public-methods,too-many-return-statements,too-few-public-methods,duplicate-code
-#,redefined-outer-name,too-many-locals,too-many-instance-attributes,import-error,missing-module-docstring,missing-function-docstring'
 
 [pylint:MASTER]
 ignore-patterns=test_(.)*\.py,bindthem.py,version.py
-ignore=tests,amg_core
+ignore=tests,amg_core,gallery,classical,vis,relaxation
